pyproject.toml
[project]
name = "good-start"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.30",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "python-frontmatter>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=9.10.0",
    "pytest>=9.0.2",
]
